[CASCL-1386] Move ExtractEC2InstanceID to common/aws so every caller can dedup
`ExtractEC2InstanceID` and `LabelEKSNodegroup` were introduced in
`common/clusterinfo` so the evict package could reuse them, but two more
copies of the same `aws:///[^/]+/(i-[0-9a-f]+)$` regex remained inline:
in `common/clusterinfo/classify.go:197` and `common/karpenter/fromnodes.go:65`.

`common/clusterinfo` already imports `common/karpenter` (for the Karpenter
detection helpers), so karpenter can't reach back into clusterinfo without
producing an import cycle. Move the helper to a neutral home —
`common/aws/node.go` — that both can import. The unit test moves with the
function.

Net result: one regex definition, four call sites, three packages dedup'd.
